refactor: rename component SnackBar to Snackbar
