--- src/shared/components/Layout/SnackBar/index.tsx
@@ -1,1 +1,0 @@
-export { default } from "./SnackBar";
