+++ src/shared/components/Layout/Snackbar/Snackbar.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { useSelector } from "react-redux";
 import useSnackbar from "shared/common/hook/useSnackbar";
 import { ISnackBarState } from "shared/common/store/SnackBar/types/Snackbar.types";
-import SnackBarView from "./SnackBarView";
+import SnackBarView from "./SnackbarView";
 
 const THREE_SECONDS_AUTO_HIDE_DURATION = 3000;
