+++ src/shared/components/Layout/Snackbar/SnackbarView.tsx
