+++ src/shared/components/Layout/Snackbar/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./Snackbar";
