+++ src/shared/components/Layout/Snackbar/types/SnackBar.component.ts
